Re-type delta t, explicitly size error args to get_pid.
git-svn-id: https://arducopter.googlecode.com/svn/trunk@936 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ libraries/PID/PID.cpp
@@ -9,7 +9,7 @@
 #include "PID.h"
 
 long
-PID::get_pid(long error, long dt, float scaler)
+PID::get_pid(int32_t error, uint16_t dt, float scaler)
 {
 	float output		= 0;
  	float delta_time	= (float)dt / 1000;
